Add remaining binaries that are built by the makefiles.
Fix compiler warning.

+++ man/mantohtml.c
@@ -513,7 +513,7 @@ main(int  argc,				/* I - Number of command-line args */
 	float amount = 3.0;		/* Indentation */
 
         if (line[3])
-          amount = atof(line + 4);
+          amount = (float)atof(line + 4);
 
 	fputs(end_fonts[font], outfile);
 	font = 0;
@@ -559,7 +559,7 @@ main(int  argc,				/* I - Number of command-line args */
 	float amount = 3.0;		/* Indentation */
 
         if (line[3])
-          amount = atof(line + 4);
+          amount = (float)atof(line + 4);
 
 	fputs(end_fonts[font], outfile);
 	font = 0;
@@ -592,7 +592,7 @@ main(int  argc,				/* I - Number of command-line args */
 	float amount = 3.0;		/* Indentation */
 
         if (line[3])
-          amount = atof(line + 4);
+          amount = (float)atof(line + 4);
 
 	fputs(end_fonts[font], outfile);
 	font = 0;
@@ -680,7 +680,7 @@ main(int  argc,				/* I - Number of command-line args */
           lineptr ++;
 
         if (isdigit(*lineptr & 255))
-          amount = atof(lineptr);
+          amount = (float)atof(lineptr);
 
         if (newlist && list && strcmp(newlist, list))
         {
